AB#2780 - change address init page
